Moved all files into cpp/ directory.
Settings `includes` allows users to include headers from this project using
angle brackets. Setting includes = ["."] makes the point_one/ directory
available like `#include <point_one/...>`. Doing it in the workspace directory
technically works and builds without error when the project is imported in the
WORKSPACE of another project, but it also is exposing _any_ subdirectories in
the workspace directory, not just point_one/.

When building the library locally as `bazel build //:messages`, however, Bazel
considers "." an error. To get around this, we move everything into a
subdirectory.