Refactored to use top-level BUILD file.
Unfortunately, Bazel's options for setting the include path are not great.
Ideally, we want users to be able to include files using <> (this is our own
preferred way to use third-party libraries; we use "" exclusively for our code).
Additionally, we want to use absolute paths for includes.

In order to add files to the include "system" search path to enable <>, the
`cc_library` must set `includes`. You cannot set `includes` to the repo root
directory, so all source files must be stored in a subdirectory (`src/` in our
case).

Unfortunately, you also cannot use `..` to set `includes` for a `cc_library`
defined in a child directory to point to a parent. As a result, absolute paths
must include the `src/` prefix, which then won't work in the user's code. Even
if you _could_ use relative paths, at best it would mean you'd have to set
`includes` on every single cc_`library`. Since `includes` is not transitive,
the absolute paths for child libraries don't work even if you start the Bazel
build from a top-level target that does set `includes`.

Bazel also does not allow the user to set `includes` themselves when they import
a library in their `WORKSPACE` file. In other words, if the library doesn't
explicitly support <> includes, there's nothing the user can do about it other
than manually setting compiler options.

We decided that the best option, which allows for both <> and absolute paths, is
to have a single `BUILD` file at the top of the repo that defines _all_ targets
and also sets `includes`. This is less Bazel-y than allowing different
directories to declare their own libraries and dependencies, but at least it
works. Since this repo is expected to be reasonably small, it should hopefully
be ok.