Changed CRC check to early return to clarify code flow and added warning options.
This works the same way as the previous version, but is a little more
straight-forward to read for customers.

It also adds constructor controls for gap and error warnings in favor of the
previous print_seq_skip_warnings() function, which was inconsistent with the
warn_on_unrecognized option.

The warn_on_unrecognized behavior now warns _after_ CRC validation to avoid a
flood of false positives.