Corrected application path injection.
We want the current directory to be inserted at the front of the path, not the
end. If it is inserted at the end, if the user has a different version of
FusionEngine Client installed as a pip package in their system Python
installation, that will get used instead of this one and may cause unexpected
behavior.