Fixed crash when generating an index for a file starting with a pose message.
The FileReader class will read and cache one message on startup in order to
establish P1 time t0. When it does, it does not generate an index file.

If that message happens to be a Pose message, calling generate_index()
immediately after will return the _cached_ result and not actually produce the
index file. We need to force it to ignore the cached value.