Added ExtendedBooleanAction (vs tri-state).
Tri-state actions should be used only where we intentionally want None to
indicate that the user did not specify anything at all. If we want the value
specifier behavior of the tri-state action but want it to default to False if
unspecified, we can either manually set default=False, or use this new extended
action instead.