Don't cache initial t0 reads.
Doing this can cause max_messages to fail unexpectedly. Say the log begins with
a non-P1 message. The initial P1 t0 read will not return that message:
  EventNotificationMessage (system time 0.0)
  PoseMessage (P1 time 1.0)  <-- Initial read returns this

Later, if we do a read with max_messages=1 and do not also specify a set of
message types, we would expect the read to return the event, but since the
previous read parameters match the current read (all message types, max=1), it
will unexpectedly return the cached data instead.