Destruct the index builder when finished.
If we don't do this, the next rewind will reset the builder and
attempt to generate the index a second time. If that second time includes
filters (e.g., pose messages only), the resulting index will only include a
subset of the data since we now have an in-memory index to use and don't
need to actually query the file.