Print logged commands/responses in the event table.
When the device responds to a command, generally the response message will be
sent back to the host on the same interface and that interface will _not_
receive a COMMAND_RESPONSE EventNotificationMessage. This allows us to display
those responses if the data log was captured on the same interface that the
commands were issued.