Fixed DLL export error on FusionEngineFramer for Windows builds.
With MSVC, if a class is marked as __declspec(dllexport), it will fail to build
(compiler warning, but we treat warnings as errors) if any of its non-POD
members are not also marked as dllexport. STL containers like std::function are
not exported.

Microsoft's recommendation is to mark individual public functions as dllexport
instead of the entire class.

See
https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-1-c4251.