Fixed NamedTupleAdapter handling of already-translated values.
It the input to NamedTupleAdapter is not a construct Containers object, but is
something that has already been transformed from a Containers object (e.g.,
using AutoEnum to transform a Containers result to an enum), we need to use it
directly as input to the NamedTuple. It will not have a values() function.